moved contract to document folder
