+++ apps/crm/crm/src/app/crm/pages/waterfall/document/document.module.ts
@@ -14,7 +14,7 @@ import { RightHolderNamePipeModule } from '@blockframes/waterfall/pipes/righthol
 import { ExpenseTypesFormModule } from '@blockframes/waterfall/components/forms/expense-types-form/form.module';
 import { VersionSelectorModule } from '@blockframes/waterfall/components/version/version-selector/version-selector.module';
 import { InterestTableModule } from '@blockframes/waterfall/components/interests-table/interest-table.module';
-import { ContractMainInfoModule } from '@blockframes/waterfall/components/contract/contract-main-info/contract-main-info.module';
+import { ContractMainInfoModule } from '@blockframes/waterfall/components/document/contract/contract-main-info/contract-main-info.module';
 import { ContractPipeModule } from '@blockframes/contract/contract/pipe';
 
 // Material
