+++ libs/waterfall/src/lib/components/document/contract/contract-list/contract-list.component.scss
