+++ libs/waterfall/src/lib/components/document/contract/contract-main-info/contract-main-info.component.scss
